requirements.txt
Django==4.2.8
djangorestframework==3.14.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
PyGithub==2.1.1
mailgun>=0.1
django-celery-beat==2.5.0
python-dateutil==2.8.2
